// fe/src/lib/api/generated/model/familyMemberPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { AnnualIncomeStatusValue } from './annualIncomeStatusValue';
import type { FamilyMemberKindValue } from './familyMemberKindValue';
import type { OccupationCategoryValue } from './occupationCategoryValue';

export interface FamilyMemberPayload {
  annual_income_previous_year?: string | null;
  annual_income_status?: AnnualIncomeStatusValue | null;
  business_address?: string;
  business_telephone?: string;
  contact_number?: string;
  current_address?: string;
  current_religion?: string;
  date_of_birth?: string | null;
  educational_attainment?: string;
  email_address?: string;
  kind: FamilyMemberKindValue;
  languages_spoken?: string;
  name?: string;
  occupation?: string;
  occupation_category?: OccupationCategoryValue | null;
  permanent_address?: string;
  place_of_birth?: string;
  religion_raised_with?: string;
}